Hkrb bump
There is weirdness here on WiiU.
These files seem to be rounded to the nearest 32, in contravention of the rest of the WiiU file behavior.
Also, the 40 byte parse size seems to be erroneous?
So the calc is always 72 bytes higher than necessary.
Leaving the WiiU calc untouched. The upshot is that it'll always work.